+++ core/src/localdb/migrations/20250801095832_add_short_ids.up.sql
@@ -1,0 +1,7 @@
+-- Add up migration script here
+
+CREATE TABLE short_ids (
+    short_id INTEGER PRIMARY KEY AUTOINCREMENT,
+    uid      TEXT UNIQUE NOT NULL,
+    kind     TEXT NOT NULL
+);