+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `aim done` support multiple todos now
 - sqlite database is persistent now
 
+### Changed
+
+- `short_id` has been move into sqlite, all events or todos will be assign a new id
+
 ### Deprecated
 
 - `aim undo` is a deprecated shortcut of `aim todo undo` now, will be remove in v0.4.0